refactor: move zeroex quote function
