+++ src/quote/zeroEx/function.test.ts
@@ -9,7 +9,7 @@ import {
 } from 'constants/tokens'
 import { ZeroExApi } from 'utils/0x'
 import { wei } from 'utils/numbers'
-import { getFlashMintZeroExQuote, getRequiredComponents } from '.'
+import { getFlashMintZeroExQuote, getRequiredComponents } from './function'
 
 const index0xApiBaseUrl = process.env.INDEX_0X_API
 const provider = new JsonRpcProvider(process.env.MAINNET_ALCHEMY_API, 1)
